Remove Paths type or tree
Breaking changes:
- Rename Byte to Tree
- Rename Walk to WalkFrom
- Rename WalkPath to WalkTo
- Remove Iterator interface
- Rename bytesTreeIter to Iterator
- Rename BytesIterator to Stepper

Other changes:
- Modify benchmarks to indicate performance of each operation
- Update README and examples
- Download benchmark data from Itnernet
